Finished comm in a new window back-end
@DaveSauce I didn’t go into less to see why the documents and team is
not shown in the new window (it’s because on the classic dashboard they
are initialized closed and then are opened with JS). I believe you’ll
rehaul some css if you’ll use it for the mobile part also so i’ve left
it up to you if you don’t mind.

You can get into the comm area with /comm or using the open in a popup
button.

All features should work, best thing for it is that the code it’s the
same with the one in the dashboard so all logic is the same.

Will you take over this branch @DaveSauce ? I’ll make another for the
other things i have